requirements.txt
napari-plugin-engine>=0.1.4
numpy
napari-tools-menu
napari-time-slicer
laptrack>=0.14.0
napari_skimage_regionprops>=0.9.0
pandas
